Refactor taxa list into component
